vectorazul-aws-config/Funciones/requirements.txt
os
logging
jsonpickle
boto3
aws_xray_sdk.core